milestone3/yizhen/src/workflow/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "DnD Model Training"
requires-python = ">=3.10"

dependencies = [
    "tensorflow==2.15.0",
    "google-cloud-aiplatform==1.120.0",
]